refactor(proxy): organize module structure with logical sub-modules
- Create focused sub-modules: http, paths, storage for better organization
- Group related functionality: HTTP types/utilities, path implementations, storage
- Simplify imports with cleaner public API surface
- Remove module nesting complexity while maintaining encapsulation
- Update all internal imports to use new module paths
- Fix benchmark imports to use public API modules
- Allow dead code for future configuration-based middleware setup

This improves code discoverability and makes the module structure
more intuitive for users of the proxy module.
